implementacija rezervacije na frontendu

+++ backend/src/main/java/com/project/Rentingaccommodation/controller/ReservationController.java
@@ -78,20 +78,21 @@ public class ReservationController {
 					return new ResponseEntity<>("User doesnt exist.", HttpStatus.FORBIDDEN);
 				}
 				
-				Pattern pattern = Pattern.compile("^\\d{4}\\-(0?[1-9]|1[012])\\-(0?[1-9]|[12][0-9]|3[01])$");
+//				Pattern pattern = Pattern.compile("^\\d{4}\\-(0?[1-9]|1[012])\\-(0?[1-9]|[12][0-9]|3[01])$");
+				Pattern pattern = Pattern.compile("^[0-3]?[0-9]/[0-3]?[0-9]/(?:[0-9]{2})?[0-9]{2}$");
 				Matcher startDateMatcher = pattern.matcher(reservation.getStartDate());
 				Matcher endDateMatcher = pattern.matcher(reservation.getEndDate());
 				
 				// Check date formats.
 				if (!startDateMatcher.find()) {
-					return new ResponseEntity<>("Start date must be in format yyyy-MM-dd.", HttpStatus.FORBIDDEN);
+					return new ResponseEntity<>("Start date must be in format dd/MM/yyyy.", HttpStatus.FORBIDDEN);
 				}
 				if (!endDateMatcher.find()) {
-					return new ResponseEntity<>("End date must be in format yyyy-MM-dd.", HttpStatus.FORBIDDEN);
+					return new ResponseEntity<>("End date must be in format dd/MM/yyyy.", HttpStatus.FORBIDDEN);
 				}
 				
 				// Check if dates are past dates.
-				SimpleDateFormat dateFormatter = new SimpleDateFormat("yyyy-MM-dd");
+				SimpleDateFormat dateFormatter = new SimpleDateFormat("dd/MM/yyyy");
 				Date startDate = dateFormatter.parse(reservation.getStartDate());
 				Date endDate = dateFormatter.parse(reservation.getEndDate());
 				if (startDate.before(new Date()) || endDate.before(new Date())) {
@@ -130,11 +131,22 @@ public class ReservationController {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-	public ResponseEntity<Object> deleteReservation(@PathVariable Long id) {
-		Reservation reservation = service.delete(id);
-		if (reservation == null) {
-			return new ResponseEntity<>("Reservation not found.", HttpStatus.NOT_FOUND);
+	public ResponseEntity<Object> deleteReservation(@PathVariable Long id, @RequestHeader("Authorization") String authHeader) {
+		try {
+			String token = authHeader.split(" ")[1].trim();
+			JwtUser jwtUser = jwtValidator.validate(token);
+			if (jwtUser != null) {
+				Reservation reservation = service.delete(id);
+				if (reservation == null) {
+					return new ResponseEntity<>("Reservation not found.", HttpStatus.NOT_FOUND);
+				} else {
+					return new ResponseEntity<>(reservation, HttpStatus.OK);
+				}
+			} else {
+				return new ResponseEntity<>("User with this email doesn't exist.", HttpStatus.NOT_FOUND);
+			}
+		} catch (Exception ex) {
+			return new ResponseEntity<>("oken not provided.", HttpStatus.NOT_FOUND);
 		}
-		return new ResponseEntity<>(reservation, HttpStatus.OK);
 	}
 }
